requirements.txt
Flask==2.1.3
numpy==1.22.3
opencv_python_headless==4.6.0.66
Pillow==9.2.0
gunicorn==20.1.0
torch==1.12.0+cu113
torchvision==0.13.0+cu113
